requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
pydantic==2.5.3
pydantic-settings==2.1.0
folium
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
geoalchemy2==0.14.3

xarray==2024.1.0
netCDF4==1.6.5
pandas==2.2.0
numpy==1.26.3
pyarrow==15.0.0

chromadb==0.4.22
faiss-cpu==1.7.4
sentence-transformers==2.3.1

groq==0.11.0
openai==1.10.0
anthropic==0.18.0
langchain==0.1.6
langchain-openai==0.0.5
tiktoken==0.5.2

plotly==5.18.0
folium==0.15.1
streamlit-folium==0.16.0
matplotlib==3.8.2
seaborn==0.13.1

shapely==2.0.2
pyproj==3.6.1
geopy==2.4.1

python-dotenv==1.0.1
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
redis==5.0.1
celery==5.3.6

loguru==0.7.2
prometheus-client==0.19.0

pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
